requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.12.1
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core>=2.25.0
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.47.0
googleapis-common-protos==1.74.0
groq==1.0.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.20.0
numpy>=2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==21.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
pyparsing==3.3.2
pypdfium2==5.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
